Add option to parse ARP table
Default behavior is to treat all the ARP table entries that DD-WRT returns from /Status_Lan.live.asp as connected. This reflects the behavior of arptable at the command line so is a logical default. However, some versions of DD-WRT add stale clients to the returned ARP table, choosing to show them as inactive by reporting zero connections instead of trimming them altogether. This change adds an option to parse the returned ARP table and trim off stale zero-connection clients, improving the function of this component when used for presence detection.